In LiquidTransportParams we added pointers to LiquidTranInteraction objects for each of the transport properties: viscosity, thermalCond, speciesDiffusivity,  electCond, hydroRadius.  This should allow us to remove the members like visc_Eij, visc_Sij, thermalCond_Aij, diff_Dij.  Also should be able to remove LiquidTranMixingModel model_* members.  This removal is pending some testing.
class LiquidTranInteraction has been further developed along with
subclasses.  The list of subclasses that should be in reasonable shape
is (all should be tested)
  class LTI_Solvent;
  class LTI_MoleFracs;
  class LTI_MassFracs;
  class LTI_Log_MoleFracs;
  class LTI_Pairwise_Interaction;
  class LTI_StokesEinstein;

The constructor takes only the TransportPropertyList enum.  The init:
    virtual void init( const XML_Node &compModelNode = 0, 
	  thermo_t* thermo = 0 );			  
takes the thermo object along with the XML node
<compositionDependence> which it parses.  The major methods for
LiquidTranInteraction are:
    virtual void setParameters( LiquidTransportParams& trParam ) 
that sets additional things required from trParam for some
subclasses.  
    virtual doublereal getMixTransProp( doublereal* speciesValues, doublereal *weightSpecies = 0 ) 
    virtual doublereal getMixTransProp( std::vector<LTPspecies*> LTPptrs )
that get a mixture averaged transport property.
    virtual DenseMatrix getMatrixTransProp( doublereal* speciesValues = 0 ) { return m_Dij; }
that get a matrix of interaction transport properties.

class LiquidTranInteraction holds members for various types of
interactions.  Some documentation of the purpose of each of these is
still required.