Proof-of-concept for a Python interface written using Cython
This module is compatible with both Python 2 and Python 3. Unlike the existing
Python module, this module directly utilizes the Cantera C++ interface,
bypassing the "clib" compatibility layer, as well as all of the direct use of
the Python C API.

Currently, this contains just enough to instantiate a ThermoPhase object from an
XML input file.